Moved QueryExplanation to Answer package
